Update angular-scroll-animate.js
I had to add this $timeout in order to wait for my elements to be ready. Without this $timeout, sometimes elements that were already on screen and had the scroll directive wouldn't appear, since their position were not right since they were not ready.

In my case, to reproduce this you simply have an element with the directive and it needs to bee fully appearing on screen. I don't know if this is the best place for the $timeout to be, but it worked for me.

I hope it helps.